refactor(landing): landing renamed to home and skeleton error fix
